requirements/base_requirements.txt
Django==1.8.5
unicode-slugify==0.1.3
lazy==1.2
django-filter==0.11.0
django-braces==1.8.1
bidict==0.9.0.post1